docshell/test/mochitest/test_compressed_multipart.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1600211

Loads a document that is served as multipart/x-mixed-replace as well as gzip compressed.
Checks that we correctly decompress and display it (via running JS within the document to notify us).
-->
<head>
  <title>Test for Bug 1600211</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1600211">Mozilla Bug 1600211</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 1600211 */
SimpleTest.waitForExplicitFinish();

var w;

function finishTest() {
  is(w.document.documentElement.textContent, "opener.finishTest();");
  is(w.document.documentElement.innerHTML, "<head><script>opener.finishTest();</" +
                                           "script></head>");
  w.close();
  SimpleTest.finish();
}

w = window.open("file_compressed_multipart");

</script>
</pre>
</body>
</html>
